fix `NoUncheckedIndexedAccess` and move compilerOptions types to their own file
