+++ src/functions/Array.ts
@@ -11,7 +11,7 @@ import {
     TupleOfUpToButNotIncluding,
 } from '../types/Array'
 import { Enumerate } from '../types/Number'
-import { FindResult } from '../types/internal'
+import { FindResult } from '../types/_internal'
 import { Keys } from '../types/misc'
 import { isNullOrUndefined } from './misc'
 import { castArray as lodashCastArray, orderBy } from 'lodash'
